% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{addElem_dLinkedList}
\alias{addElem_dLinkedList}
\title{dLinkedList: Add Element}
\usage{
addElem_dLinkedList(new, prev, nex)
}
\arguments{
\item{llist}{is a doubly linked list}
}
\description{
Adds element to end of a doubly linked list
}
