% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashMap-Class.R
\name{assign_HashMap}
\alias{assign_HashMap}
\title{MASH-CPP: HashMap Assign a Key-Value Pair to the Hash Table}
\usage{
assign_HashMap(key, value)
}
\arguments{
\item{key}{character key}

\item{value}{a value to be assigned}
}
\description{
Assign a key-value pair to \code{private$storage}.
}
