% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEBUG-MASHcpp.R
\name{DEBUG.MASHCPP}
\alias{DEBUG.MASHCPP}
\title{MASH-CPP: Debug Finalizers for C++ Objects}
\usage{
DEBUG.MASHCPP(overwrite = TRUE)
}
\description{
Print when objects are garbage collected. For debugging purposes only.
}
