% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{delElement}
\alias{delElement}
\title{dLinkedList: Check if Value Exists}
\usage{
delElement(llist, pos = NULL)
}
\arguments{
\item{llist}{is an instance of the doubly linked list class}
}
\description{
Does a value exist in \code{private$storage} or not.
}
