% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashMap-Utility.R
\name{eapplyInvisible}
\alias{eapplyInvisible}
\title{MASH-CPP: eapply with Invisisble Return}
\usage{
eapplyInvisible(envir)
}
\arguments{
\item{envir}{an environment to iterate over}
}
\description{
write me
}
