% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\docType{class}
\name{ELPool}
\alias{ELPool}
\title{ELPool Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
ELPool(alpha_new = 0.1, gamma_new = 0.1, psi_new = 0.01, sigma_new = 1)
}
\description{
LPool
}
\section{\strong{Description}}{

Generate a single ELPool, a simplified model of aquatic ecology which only includes egg and a single larval stage. See \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4357799/} and \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3737150/} for details on model formulation. The daily difference equations are thus formulated as \deqn{L_{t+1}=E_{t}-(\alpha +\gamma +\psi L_{t}^{\sigma})L_{t}} where aquatic stages mature at the constant per-capita rate \eqn{\alpha}. Mortality is given by the per-capita rate \eqn{\alpha +\gamma +\psi L_{t}^{\sigma}}, where \eqn{\gamma} describes all sources of density independent mortality and and the power-law function \eqn{\psi L_{t}^{\sigma}} describes mortality rates as a function of mean density and \eqn{\psi} represents increase in per-capita mortality in response to crowding. For \eqn{\sigma = 1} the relationship gives mean crowding, which is analogous to the classical first-order description of density dependence as described by the logistic growth equation.
}

\section{\strong{Constructor}}{

* alpha_new = 0.1: daily mautration rate of aquatic stages (1/alpha is expected duration of aquatic life stages)
* psi_new = 0.01: density dependent mortality; increase in per-capita mortality in response to crowding (unit given as 1/L)
* gamma_new = 0.1: daily density independent mortality rate
* sigma_new = 1: scaling relationship of density dependence in response to crowding
}

\section{\strong{Methods}}{

\itemize{
\item oneStep: run daily difference equation
\item get_psi:
\item set_psi:
\item add_ImagoQ:
\item track_ImagoQ:
\item get_N:
\item set_N:
\item get_ImagoQ:
\item get_ImagoQTime:
}
}

\section{\strong{Fields}}{

\itemize{
\item a field:
}
}

\keyword{R6}
\keyword{class}
