% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{exists_dLinkedList}
\alias{exists_dLinkedList}
\title{dLinkedList: Check if Value Exists}
\usage{
exists_dLinkedList(llist, value, exists = FALSE)
}
\arguments{
\item{llist}{is an instance of the doubly linked list class}
}
\description{
Does a value exist in \code{private$storage} or not.
}
