% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{HistoryTravel}
\alias{HistoryTravel}
\title{Travel History}
\usage{
HistoryTravel(N = 100)
}
\arguments{
\item{N}{reserve memory for \code{N} number of events}
}
\description{
A class for recording travel histories comprised of event times (double precision floats) and locations (integers).
}
\section{Methods}{

\itemize{
\item track_travel: record an event from arguments 'tEvent' (double) and 'event' (string)
\item get_travelHistory: return history as a named list
}
}

\section{Fields}{

\itemize{
\item tTravel: \code{std::vector<double>}
\item locationH: \code{std::vector<int>}
}
}

