% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{humanPfMOI}
\alias{humanPfMOI}
\title{PfMOI Human Stage Pathogen Class}
\usage{
humanPfMOI(b_init = 0.55, c_init = 0.15,
  chemoprophylaxis_init = FALSE)
}
\arguments{
\item{_PfID}{PfID for this clonal infection}

\item{_tInf}{time of successful mosquito to human transmission event for this clonal infection}

\item{_MOI}{initial multiplicity of infection}

\item{_b}{mosquito to human transmission efficiency}

\item{_c}{human to mosquito transmission efficiency}

\item{_chemoprophylaxis}{chemoprophylaxis protection status}
}
\description{
A class for storing parameters and state variables of human-stage PfMOI infections.
}
\section{Methods}{

\itemize{
\item get_PfID: get vector of current infections by PfID
\item get_tInf: get vector of current infections by time of transmisson event
\item get_MOI: get current multiplicity of infection
\item get_b: get mosquito to human transmission efficiency
\item set_b: set mosquito to human transmission efficiency
\item get_c: get human to mosquito transmission efficiency
\item set_c: set human to mosquito transmission efficiency
\item get_chemoprophylaxis: is human currently protected by chemoprophylaxis
\item set_chemoprophylaxis: set protection by chemoprophylaxis
\item add_Infection: add a new infection
\item clear_Infection: completely clear the infection associated with PfID
\item clear_Infections: clear all current infections
\item get_Infection: get all current infections
}
}

\section{Fields}{

\itemize{
\item PfID: std::vector<int>
\item tInf: std::vector<double>
\item MOI: int
\item b: double
\item c: double
\item chemoprophylaxis: bool
}
}

