% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{isEmpty_dLinkedList}
\alias{isEmpty_dLinkedList}
\title{dLinkedList: Check Emptiness}
\usage{
isEmpty_dLinkedList(llist)
}
\arguments{
\item{llist}{is a doubly linked list}
}
\description{
Checks if linked list is empty
}
