% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{makeNode_dLinkedList}
\alias{makeNode_dLinkedList}
\title{dLinkedList: Node Creation}
\usage{
makeNode_dLinkedList(val, prevnode = NULL, nextnode = NULL)
}
\arguments{
\item{val}{is a primitive type, prevnode and node are either NULL or other doubly linked lists}
}
\description{
Initializes a double linked list node
}
