% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{MatingQ}
\alias{MatingQ}
\title{MICRO Mating Queue}
\usage{
MatingQ()
}
\description{
Initialize a mating queue for a \code{\link[MASHmicro]{MatingSite}} object.
}
\section{Methods}{

\itemize{
\item get_N: write about me!
\item add_male2Q: write about me!
\item get_MatingQ: write about me!
\item clear_MatingQ: write about me!
}
}

\section{Fields}{

\itemize{
\item a field: write about me!
}
}

