% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{MosquitoMaleHistory}
\alias{MosquitoMaleHistory}
\title{Mosquito Male History Class}
\usage{
MosquitoMaleHistory()
}
\description{
Class to store male mosquito histories
}
\section{Methods}{

\itemize{
\item a method: write me!
}
}

