% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{mosquitoPfMOI}
\alias{mosquitoPfMOI}
\title{PfMOI Mosquito Stage Pathogen Class}
\usage{
mosquitoPfMOI()
}
\description{
A class for storing parameters and state variables of mosquito-stage PfMOI infections.
}
\section{Methods}{

\itemize{
\item get_PfID:
\item push_PfID:
\item get_tInf:
\item push_tInf:
\item get_MOI:
\item set_MOI:
\item get_damID:
\item set_damID:
\item get_sireID:
\item set_sireID:
}
}

\section{Fields}{

\itemize{
\item a field: talk about me
}
}

