% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashMap-Class.R
\name{rm_HashMap}
\alias{rm_HashMap}
\title{HashMap: Value Removal Method}
\usage{
rm_HashMap(key)
}
\arguments{
\item{key}{a character key; remove that value from the storage hash table.}
}
\description{
Remove a value from \code{private$storage} given a key.
}
