% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashMap-Class.R
\name{rmAll_HashMap}
\alias{rmAll_HashMap}
\title{HashMap: Remove all Values}
\usage{
rmAll_HashMap()
}
\arguments{
\item{key}{a character key; remove that value from the storage hash table.}
}
\description{
Remove a value from \code{private$storage} given a key.
}
