% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedList-Class.R
\name{size_dLinkedList}
\alias{size_dLinkedList}
\title{dLinkedList: Return Size of List}
\usage{
size_dLinkedList(llist, size = 0)
}
\arguments{
\item{llist}{is a doubly linked list}
}
\description{
Returns the total number of nodes in the list
}
