% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashMap-Class.R
\docType{data}
\name{HashMap}
\alias{HashMap}
\title{MASH-CPP: HashMap Class Definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HashMap
}
\description{
Provide a R6 implementation of a hash table to store population vectors using native R environments.
}
\section{Methods}{

\itemize{
\item rm: given a character key remove that value from the storage hash table.
\item exists: given a character key return a boolean if the value associated to that key exists in the storage hash table
}
}

\keyword{datasets}
